rearranging clustering coefficient algos; minor bugfixes; cleanup
