+++ Cargo.toml
@@ -155,3 +155,4 @@ arrow-buffer = { version = "53.2.0" }
 arrow-schema = { version = "53.2.0" }
 arrow-array = { version = "53.2.0" }
 moka = { version = "0.12.7", features = ["sync"] }
+hyperloglog = "1.0.2"
