+++ python/python/raphtory/algorithms/__init__.pyi
@@ -418,8 +418,38 @@ def local_clustering_coefficient(g: GraphView, v: InputNode):
         float : the local clustering coefficient of node v in g.
     """
 
-def local_clustering_coefficient_batch_intersection(graph, v): ...
-def local_clustering_coefficient_batch_path(graph, v): ...
+def local_clustering_coefficient_batch_intersection(
+    graph, v: List[InputNode] | InputNode | None
+):
+    """
+    Local clustering coefficient (batch, intersection) - measures the degree to which nodes in a graph tend to cluster together.
+    Uses cached sets of neighbours to handle triangle-counting.
+    The proportion of pairs of neighbours of a node who are themselves connected.
+
+    Arguments:
+        g (GraphView) : Raphtory graph, can be directed or undirected but will be treated as undirected.
+        v (List[InputNode] | InputNode | None): list of node ids or names, if empty, calculates LCC of all nodes
+
+    Returns:
+        AlgorithmResult : AlgorithmResult mapping vertices to local clustering coefficient.
+    """
+
+def local_clustering_coefficient_batch_path(
+    graph, v: List[InputNode] | InputNode | None
+):
+    """
+    Local clustering coefficient - measures the degree to which nodes in a graph tend to cluster together.
+    Uses path-counting for its triangle-counting step.
+    The proportion of pairs of neighbours of a node who are themselves connected.
+
+    Arguments:
+        g (GraphView) : Raphtory graph, can be directed or undirected but will be treated as undirected.
+        v (List[InputNode] | InputNode | None): list of node ids or names, if empty, calculates LCC of all nodes
+
+    Returns:
+        AlgorithmResult : AlgorithmResult mapping vertices to local clustering coefficient.
+    """
+
 def local_temporal_three_node_motifs(g: GraphView, delta: int):
     """
     Computes the number of each type of motif that each node participates in. See global_temporal_three_node_motifs for a summary of the motifs involved.
