+++ raphtory/src/algorithms/metrics/mod.rs
@@ -2,5 +2,4 @@ pub mod balance;
 pub mod clustering_coefficient;
 pub mod degree;
 pub mod directed_graph_density;
-pub mod local_clustering_coefficient;
 pub mod reciprocity;
